server: move server-side persistence types into server module
